+++ metric/humanparsing/mhp_extension/detectron2/tests/modeling/test_fast_rcnn.py
@@ -27,8 +27,8 @@ class FastRCNNTest(unittest.TestCase):
         feature_pooled = torch.rand(2, box_head_output_size)
         predictions = box_predictor(feature_pooled)
 
-        proposal_boxes = torch.tensor([[0.8, 1.1, 3.2, 2.8], [2.3, 2.5, 7, 8]], dtype=torch.float32)
-        gt_boxes = torch.tensor([[1, 1, 3, 3], [2, 2, 6, 6]], dtype=torch.float32)
+        proposal_boxes = torch.tensor([[0.8, 1.1, 3.2, 2.8], [2.3, 2.5, 7, 8]], dtype=torch.float16)
+        gt_boxes = torch.tensor([[1, 1, 3, 3], [2, 2, 6, 6]], dtype=torch.float16)
         proposal = Instances((10, 10))
         proposal.proposal_boxes = Boxes(proposal_boxes)
         proposal.gt_boxes = Boxes(gt_boxes)
@@ -79,9 +79,9 @@ class FastRCNNTest(unittest.TestCase):
         feature_pooled = torch.rand(2, box_head_output_size)
         predictions = box_predictor(feature_pooled)
         proposal_boxes = torch.tensor(
-            [[2, 1.95, 2.4, 1.7, 0], [4.65, 5.25, 4.7, 5.5, 0]], dtype=torch.float32
+            [[2, 1.95, 2.4, 1.7, 0], [4.65, 5.25, 4.7, 5.5, 0]], dtype=torch.float16
         )
-        gt_boxes = torch.tensor([[2, 2, 2, 2, 0], [4, 4, 4, 4, 0]], dtype=torch.float32)
+        gt_boxes = torch.tensor([[2, 2, 2, 2, 0], [4, 4, 4, 4, 0]], dtype=torch.float16)
         proposal = Instances((10, 10))
         proposal.proposal_boxes = RotatedBoxes(proposal_boxes)
         proposal.gt_boxes = RotatedBoxes(gt_boxes)
